leader_board/requirements.txt
firebase-admin==6.6.0
Flask==3.0.3
python-dotenv==1.0.1
Flask-Cors==5.0.0